telegram-bot-motorcycle-technical-data/bot_requirements.txt
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
loguru==0.7.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
peewee==3.17.1
platformdirs==4.2.0
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
requests==2.31.0
tomli==2.0.1
typing_extensions==4.10.0rc1
urllib3==2.2.1
